Refactor: move all quantization-related code to `srt/layer/quantization` (#7989)
